ITERATION_3_Solution/TextToSpeechModule/requirements.txt
-e ./TTS/
torchaudio
pydub
ffmpeg-normalize==1.21.0
numpy==1.20